pytest: convert many tests to use inline_plugin helper.
Reduces the number of misc test plugins.
