+++ tests/test_misc.py
@@ -999,12 +999,12 @@ def test_malformed_rpc(node_factory):
 def test_valid_json_cli(node_factory):
     """Make sure lightning-cli passes valid json values, so that rust and python plugins
     don't crash."""
-    l1 = node_factory.get_node(
-        options={
-            "log-level": "io",
-            "plugin": os.path.join(os.getcwd(), "tests/plugins/validatejson.py"),
-        }
-    )
+    def setup(plugin):
+        @plugin.method('validate-json-rpc')
+        def validate_json_rpc(plugin, *args, **kwargs):
+            return {}
+
+    l1 = node_factory.get_node(options={"log-level": "io"}, inline_plugin=setup)
     # If passed as a literal number rust's serde_json::from_str will fail as the
     # leading zero makes it invalid for an integer.
     nodeid = "030000000000000000000000000000000000000000000000000000000000000001"
@@ -1203,7 +1203,18 @@ def test_cli(node_factory):
 
 
 def test_cli_multiline_help(node_factory):
-    l1 = node_factory.get_node(options={'plugin': os.path.join(os.getcwd(), 'tests/plugins/multiline-help.py')})
+    def setup(plugin):
+        from pyln.client import Millisatoshi
+
+        @plugin.method("helpme")
+        def helpme(plugin, msat: Millisatoshi):
+            """This is a message which consumes multiple lines and thus should
+            be well-formatted by lightning-cli help
+
+            """
+            return {'help': msat}
+
+    l1 = node_factory.get_node(inline_plugin=setup)
 
     out = subprocess.check_output(['cli/lightning-cli',
                                    '--network={}'.format(TEST_NETWORK),
@@ -1431,15 +1442,40 @@ def test_funding_reorg_private(node_factory, bitcoind):
     """
     # Rescan to detect reorg at restart and may_reconnect so channeld
     # will restart.  Reorg can cause bad gossip msg.
-    opts = {'funding-confirms': 2, 'rescan': 10, 'may_reconnect': True,
-            'allow_bad_gossip': True,
-            # gossipd send lightning update for original channel.
-            'allow_warning': True,
-            'dev-fast-reconnect': None,
-            # if it's not zeroconf, we'll terminate on reorg.
-            'plugin': os.path.join(os.getcwd(), 'tests/plugins/zeroconf-selective.py'),
-            'zeroconf_allow': 'any'}
-    l1, l2 = node_factory.line_graph(2, fundchannel=False, opts=opts)
+    def setup(plugin):
+        plugin.add_option(
+            'zeroconf_allow',
+            '03864ef025fde8fb587d989186ce6a4a186895ee44a926bfc370e2c366597a3f8f',
+            'A node_id to allow zeroconf channels from',
+        )
+        plugin.add_option(
+            'zeroconf_mindepth',
+            0,
+            'Number of confirmations to require from allowlisted peers',
+        )
+
+        @plugin.hook('openchannel')
+        def on_openchannel(openchannel, plugin, **kwargs):
+            plugin.log(repr(openchannel))
+            mindepth = int(plugin.options['zeroconf_mindepth']['value'])
+
+            if openchannel['id'] == plugin.options['zeroconf_allow']['value'] or plugin.options['zeroconf_allow']['value'] == 'any':
+                plugin.log(f"This peer is in the zeroconf allowlist, setting mindepth={mindepth}")
+                return {'result': 'continue', 'mindepth': mindepth}
+            else:
+                return {'result': 'continue'}
+
+    cli_opts = {'funding-confirms': 2, 'rescan': 10,
+                'dev-fast-reconnect': None,
+                # if it's not zeroconf, we'll terminate on reorg.
+                'zeroconf_allow': 'any'}
+    node_opts = {'may_reconnect': True,
+                 'allow_bad_gossip': True,
+                 # gossipd send lightning update for original channel.
+                 'allow_warning': True}
+    l1 = node_factory.get_node(inline_plugin=setup, options=cli_opts, **node_opts)
+    l2 = node_factory.get_node(inline_plugin=setup, options=cli_opts, **node_opts)
+    node_factory.join_nodes([l1, l2], fundchannel=False)
     l1.fundwallet(10000000)
     sync_blockheight(bitcoind, [l1])                # height 102
     bitcoind.generate_block(3)                      # heights 103-105
@@ -1477,12 +1513,36 @@ def test_funding_reorg_remote_lags(node_factory, bitcoind):
     """Nodes may disagree about short_channel_id before channel announcement
     """
     # may_reconnect so channeld will restart; bad gossip can happen due to reorg
-    opts = {'funding-confirms': 1, 'may_reconnect': True, 'allow_bad_gossip': True,
-            'allow_warning': True, 'dev-fast-reconnect': None,
-            # if it's not zeroconf, l2 will terminate on reorg.
-            'plugin': os.path.join(os.getcwd(), 'tests/plugins/zeroconf-selective.py'),
-            'zeroconf_allow': 'any'}
-    l1, l2 = node_factory.line_graph(2, fundchannel=False, opts=opts)
+    def setup(plugin):
+        plugin.add_option(
+            'zeroconf_allow',
+            '03864ef025fde8fb587d989186ce6a4a186895ee44a926bfc370e2c366597a3f8f',
+            'A node_id to allow zeroconf channels from',
+        )
+        plugin.add_option(
+            'zeroconf_mindepth',
+            0,
+            'Number of confirmations to require from allowlisted peers',
+        )
+
+        @plugin.hook('openchannel')
+        def on_openchannel(openchannel, plugin, **kwargs):
+            plugin.log(repr(openchannel))
+            mindepth = int(plugin.options['zeroconf_mindepth']['value'])
+
+            if openchannel['id'] == plugin.options['zeroconf_allow']['value'] or plugin.options['zeroconf_allow']['value'] == 'any':
+                plugin.log(f"This peer is in the zeroconf allowlist, setting mindepth={mindepth}")
+                return {'result': 'continue', 'mindepth': mindepth}
+            else:
+                return {'result': 'continue'}
+
+    cli_opts = {'funding-confirms': 1, 'dev-fast-reconnect': None,
+                # if it's not zeroconf, l2 will terminate on reorg.
+                'zeroconf_allow': 'any'}
+    node_opts = {'may_reconnect': True, 'allow_bad_gossip': True, 'allow_warning': True}
+    l1 = node_factory.get_node(inline_plugin=setup, options=cli_opts, **node_opts)
+    l2 = node_factory.get_node(inline_plugin=setup, options=cli_opts, **node_opts)
+    node_factory.join_nodes([l1, l2], fundchannel=False)
     l1.fundwallet(10000000)
     sync_blockheight(bitcoind, [l1])                # height 102
